WIP splitting up crawl.py...
